Create if statement to trigger error in notes controller functions

+++ src/notes/controller/notes.controller.js
@@ -56,9 +56,15 @@ const getNote = async (req, res) => {
 const createNote = async (req, res) => {
   const { title, content, userId } = req.body;
 
+  if (!title || !content || !userId)
+    return res.status(400).json({
+      ok: false,
+      status: 400,
+      message: 'Missing required fields',
+    });
+
   try {
     const note = await Notes.create({ title, content, userId });
-
     res.status(201).json({
       ok: true,
       status: 201,
@@ -77,12 +83,25 @@ const updateNote = async (req, res) => {
   const { id } = req.params;
   const { title, content } = req.body;
 
+  if (!id)
+    return res.status(400).json({
+      ok: false,
+      status: 400,
+      message: 'Missing required id field',
+    });
+
+  if (!title || !content)
+    return res.status(400).json({
+      ok: false,
+      status: 400,
+      message: 'Missing required title or content field',
+    });
+
   try {
     const note = await Notes.findByPk(id);
     note.title = title;
     note.content = content;
     await note.save();
-
     res.json({
       ok: true,
       status: 200,
@@ -99,6 +118,13 @@ const updateNote = async (req, res) => {
 
 const deleteNote = async (req, res) => {
   const { id } = req.params;
+
+  if (!id)
+    return res.status(400).json({
+      ok: false,
+      status: 400,
+      message: 'Missing required id field',
+    });
 
   try {
     await Notes.destroy({
